fix some compile issues (errors and warnings). sdl, ode, vulkan. fix q2bsp textures. fix some vulkan validation issues. MOVE_OTHERONLY is now an official feature (replacing MOVE_ONLYENT which is now removed, same functionality, better behaved behaviour). network up edited brushes on initial connect. still needs more work for entity editing, but should otherwise be okay for now. add sys_browserredirect console command for emscripten builds (can be used to trigger window redirections - including download requests)
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@5001 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/console.c
@@ -1600,7 +1600,7 @@ static int Con_DrawProgress(int left, int right, int y)
 {
 #ifdef RUNTIMELIGHTING
 	extern model_t *lightmodel;
-	extern int relitsurface;
+	extern long relitsurface;
 #endif
 
 	conchar_t			dlbar[1024], *chr;
